Replace lightroom-photo-tagging with portfolio-preparation
lightroom-photo-tagging did two jobs: generating the portfolio metadata
manifest and generating social media captions. The manifest half lives on
here as portfolio-preparation/, slimmed to only the fields the UI still
consumes (id, src, galleryIds, dateTaken, blurHash, width, height). The
social captioning half was ported to Python in rasppi-utils social-poster
(TravisBumgarner/rasppi-utils#35) — verified there against the original:
identical tag tree and caption output.

Also drops todolist.md (photo-curation checklist; recoverable from git
history) and updates ui/README.md to point at the new project.
