--- lightroom-photo-tagging/.env.sample
@@ -1,6 +1,0 @@
-DEBUG_MODE=false
-SOCIAL_PLATFORM=instagram | bluesky
-SOCIAL_INGEST_PATH='/Users/travisbumgarner/Desktop/cameracoffeewander_template_ingest'
-# Portfolio metadata generation reads exported .avif photos from here.
-# Must match the Lightroom "For Portfolio - Large" export dir (see metadata.ts PHOTO_DIR).
-PORTFOLIO_INGEST_PATH='/Users/travisbumgarner/Desktop/large'
