--- lightroom-photo-tagging/.prettierrc
@@ -1,12 +1,0 @@
-{
-    "printWidth": 120,
-    "tabWidth": 4,
-    "useTabs": false,
-    "semi": false,
-    "singleQuote": true,
-    "trailingComma": "es5",
-    "bracketSpacing": true,
-    "jsxBracketSameLine": false,
-    "arrowParens": "avoid",
-    "endOfLine": "auto"
-}
