--- lightroom-photo-tagging/package.json
@@ -1,35 +1,0 @@
-{
-    "name": "algorithm_explorations",
-    "version": "1.0.0",
-    "main": "index.js",
-    "license": "MIT",
-    "dependencies": {
-        "blurhash": "^2.0.5",
-        "canvas": "^2.11.2",
-        "date-fns": "^3.6.0",
-        "dotenv": "^16.4.7",
-        "exifr": "^7.1.3",
-        "runtypes": "^6.7.0",
-        "sharp": "^0.35.3",
-        "uuid": "^11.0.2",
-        "zod": "^3.24.2"
-    },
-    "devDependencies": {
-        "@types/node": "^20.12.13",
-        "@typescript-eslint/eslint-plugin": "^7.11.0",
-        "@typescript-eslint/parser": "^7.11.0",
-        "eslint": "8.56.0",
-        "eslint-config-prettier": "^9.1.0",
-        "eslint-import-resolver-typescript": "^3.6.1",
-        "eslint-plugin-import": "^2.29.1",
-        "eslint-plugin-prettier": "^5.1.3",
-        "prettier": "^3.3.2",
-        "ts-node": "^10.9.2",
-        "typescript": "^5.4.5"
-    },
-    "scripts": {
-        "test": "echo \"Error: no test specified\" && exit 1",
-        "start-social-tagging": "ts-node src/main-social-tagging.ts",
-        "start-portfolio-metadata": "ts-node src/main-portfolio-metadata.ts"
-    }
-}
