--- lightroom-photo-tagging/src/tags/FilmType.ts
@@ -1,132 +1,0 @@
-import { Tags } from '../types'
-
-const FujiPro160: Tags = {
-    general: ['#fujifilm', '#pro160'],
-    priority: [],
-    bluesky: ['#fujifilm', '#pro160'],
-}
-
-const IlfordDelta3200: Tags = {
-    general: ['#ilford', '#delta3200'],
-    priority: [],
-    bluesky: ['#ilford', '#delta3200'],
-}
-
-const Kentmere400: Tags = {
-    general: ['#kentmere400'],
-    priority: [],
-    bluesky: ['#kentmere400'],
-}
-
-const IlfordHP5: Tags = {
-    general: ['#ilford', '#hp5'],
-    priority: [],
-    bluesky: ['#ilford', '#hp5'],
-}
-
-const Kodak400TX: Tags = {
-    general: ['#kodak', '#400tx'],
-    priority: [],
-    bluesky: ['#kodak', '#400tx'],
-}
-
-const KodakUltramax400: Tags = {
-    general: ['#kodak', '#ultramax400'],
-    priority: [],
-    bluesky: ['#kodak', '#ultramax400'],
-}
-
-const Across100: Tags = {
-    general: ['#fujifilm', '#across100'],
-    priority: [],
-    bluesky: ['#fujifilm', '#across100'],
-}
-
-const Portra160: Tags = {
-    general: ['#portra160'],
-    priority: ['#MadeWithKodak'],
-    bluesky: ['#portra160', '#kodak', '#madewithkodak'],
-}
-
-const Portra400: Tags = {
-    general: ['#portra400'],
-    priority: ['#MadeWithKodak'],
-    bluesky: ['#portra400', '#kodak', '#madewithkodak'],
-}
-
-const Portra800: Tags = {
-    general: ['#portra800'],
-    priority: ['#MadeWithKodak'],
-    bluesky: ['#portra800', '#kodak', '#madewithkodak'],
-}
-
-const Gold200: Tags = {
-    general: ['#gold200', '#gold200film', '#kodakfilm', '#kodakgold200'],
-    priority: ['#MadeWithKodak'],
-    bluesky: ['#kodak', '#madewithkodak'],
-}
-
-const FujifilmXtra400: Tags = {
-    general: ['#fujifilm', '#fujifilmxtra400', '#fujifilmxtra'],
-    priority: [],
-    bluesky: ['#fujifilm'],
-}
-
-const LomographyTurquoise: Tags = {
-    general: ['#lomographyturquoise'],
-    priority: ['@lomography', '#HeyLomography'],
-    bluesky: ['#lomography'],
-}
-
-const Lomography100: Tags = {
-    general: ['#lomography100'],
-    priority: ['@lomography', '#HeyLomography'],
-    bluesky: ['#lomography'],
-}
-
-const LomographyMetro: Tags = {
-    general: ['#lomographymetropolis'],
-    priority: ['@lomography', '#HeyLomography'],
-    bluesky: ['#lomography'],
-}
-
-const LomographyPurple: Tags = {
-    general: ['#lomographypurple'],
-    priority: ['@lomography', '#HeyLomography'],
-    bluesky: ['#lomography'],
-}
-
-const KodakTMax400: Tags = {
-    general: ['#tmax400', '#kodaktmax400'],
-    priority: ['#MadeWithKodak'],
-    bluesky: ['#kodak', '#madewithkodak'],
-}
-
-const Ektar100: Tags = {
-    general: ['#ektar100', '#kodakektar100'],
-    priority: ['#MadeWithKodak'],
-    bluesky: ['#kodak', '#madewithkodak'],
-}
-
-const FilmType = {
-    Across100,
-    FujifilmXtra400,
-    KodakTMax400,
-    Gold200,
-    Lomography100,
-    LomographyMetro,
-    LomographyPurple,
-    Portra160,
-    Portra400,
-    Portra800,
-    Ektar100,
-    LomographyTurquoise,
-    IlfordDelta3200,
-    IlfordHP5,
-    KodakUltramax400,
-    FujiPro160,
-    Kentmere400,
-    Kodak400TX,
-}
-
-export default FilmType
