--- lightroom-photo-tagging/src/tags/PhotoType.ts
@@ -1,125 +1,0 @@
-import { Tags } from '../types'
-
-const Animals: Tags = {
-    general: ['#animalkingdom', '#naturewanderers', '#animallovers', '#everything_animals'],
-    priority: [],
-    bluesky: ['#animals'],
-}
-
-const Architecture: Tags = {
-    general: ['#architecture', '#architecturelovers', '#architecturephotography'],
-    priority: [],
-    bluesky: ['#architecture'],
-}
-
-const Beach: Tags = {
-    general: ['#beautifulbeaches', '#beach', '#beaches'],
-    priority: [],
-    bluesky: ['#beach'],
-}
-
-const Leaves: Tags = {
-    general: ['#leafpeeping', '#autumnleaves', '#fallcolors'],
-    priority: [],
-    bluesky: [],
-}
-
-const BlackAndWhite: Tags = {
-    general: ['#bnw', '#bnwphotography', '#bnwmood', '#bnw_captures', '#bnwphoto'],
-    priority: [],
-    bluesky: ['#blackandwhite', '#monochrome', '#bnw', '#bnw_captures'],
-}
-
-const Cars: Tags = {
-    general: ['#cars', '#carsdaily', '#carlovers', '#carswithoutlimits', '#carstagram'],
-    priority: [],
-    bluesky: [],
-}
-
-const Clouds: Tags = {
-    general: ['#clouds', '#cloudscape', '#cloudchaser', '#cloudphotography', '#cloudlovers'],
-    priority: [],
-    bluesky: ['#clouds', '#cloudscape'],
-}
-
-const Flowers: Tags = {
-    general: ['#flowers', '#flowerslovers', '#flowersphotography', '#flowersmakemehappy'],
-    priority: [],
-    bluesky: ['#flowers', '#flores'],
-}
-
-const Landscape: Tags = {
-    general: ['#landscapephotography', '#landscape_collection', '#landscapelover'],
-    priority: [],
-    bluesky: ['#landscapephotography', '#landscape'],
-}
-
-const Portrait: Tags = {
-    general: ['#portraitphotography', '#portrait'],
-    priority: ['#mundusmag'],
-    bluesky: ['#portraitphotography', '#portrait'],
-}
-
-const Selfie: Tags = {
-    general: ['#selfie', '#photographicselfportrait', '#artisticselfie'],
-    priority: [],
-    bluesky: [],
-}
-
-const Shadows: Tags = {
-    general: ['#shadows', '#shadowart'],
-    priority: ['#shadows_magazine'],
-    bluesky: ['#shadows'],
-}
-
-const Stars: Tags = {
-    general: ['#astrophotography', '#moonlovers', '#astronomy', '#nightshooters', '#nightimages'],
-    priority: ['@milkyway_nightscape', '#milkyway_nightscape_'],
-    bluesky: ['#astrophotography', '#astrophoto'],
-}
-
-const Storms: Tags = {
-    general: ['#lightening', '#stormphotography', '#stormchasers', '#stormclouds', '#thunderstorms', '#dpsp_rainshots'],
-    priority: [],
-    bluesky: ['#storm', '#rain'],
-}
-
-const Street: Tags = {
-    general: [
-        '#street_perfection',
-        '#streetphotography',
-        '#street',
-        '#streetphotographers',
-        '#streetclassics',
-        '#streetphotographer',
-    ],
-    priority: ['#streetdreamsmag'],
-    bluesky: ['#streetart', '#streetphotography'],
-}
-
-const Sunsets: Tags = {
-    general: ['#sunset', '#sunsetlovers', '#sunsets'],
-    priority: [],
-    bluesky: ['#sunset', '#colors'],
-}
-
-const PhotoType = {
-    Animals,
-    Architecture,
-    Beach,
-    BlackAndWhite,
-    Cars,
-    Clouds,
-    Flowers,
-    Landscape,
-    Leaves,
-    Portrait,
-    Selfie,
-    Stars,
-    Shadows,
-    Storms,
-    Street,
-    Sunsets,
-}
-
-export default PhotoType
