--- lightroom-photo-tagging/src/tags/Special.ts
@@ -1,30 +1,0 @@
-import { Tags } from '../types'
-
-const NatGeoYourShot: Tags = {
-    general: [],
-    priority: ['#yourshotphotographer'],
-    bluesky: [],
-}
-
-const FotoHercules: Tags = {
-    general: [],
-    priority: ['#fotohercules', '@fotohercules'],
-    bluesky: [],
-}
-
-const PhotographyTips: Tags = {
-    general: [],
-    priority: [
-        '#photographylearning',
-        '#photographytutorials',
-        '#photographytipsandtricks',
-        '#photographytips',
-        '#photographyskills',
-        '#photography101',
-    ],
-    bluesky: [],
-}
-
-const Special = { NatGeoYourShot, FotoHercules, PhotographyTips }
-
-export default Special
