--- lightroom-photo-tagging/src/tags/index.ts
@@ -1,17 +1,0 @@
-import Camera from './Camera'
-import FilmType from './FilmType'
-import PhotoType from './PhotoType'
-import SocialEvent from './SocialEvent'
-import Special from './Special'
-import Place from './place'
-
-const TAGS = {
-    Camera,
-    FilmType,
-    Place,
-    PhotoType,
-    SocialEvent,
-    Special,
-}
-
-export default TAGS
