--- lightroom-photo-tagging/src/tags/place/OtherCountries.ts
@@ -1,20 +1,0 @@
-import { Tags } from '../../types'
-
-const Iceland: Record<string, Tags> = {
-    Country: {
-        general: [
-            '#iceland',
-            '#icelandtravel',
-            '#icelandtrip',
-            '#icelandnature',
-        ],
-        priority: [],
-        bluesky: ['#iceland'],
-    },
-}
-
-const OtherCountries = {
-    Iceland,
-}
-
-export default OtherCountries
