--- lightroom-photo-tagging/todolist.md
@@ -1,134 +1,0 @@
-Photos to Go Through
-
-- [ ]  2024-10 (Let it marinate for a few months)
-- [ ]  2024-09 (Let it marinate for a few months)
-
-- [x] Move maker photos elsewhere
-- [x] Double check every month below has been looked at. 
-- [x]  2024-08
-- [x]  2024-07
-- [x]  2024-06
-- [x]  2024-05
-- [x]  2024-04
-- [x]  2024-03
-- [x]  2024-02
-- [x]  2024-01
-- [x]  2023-12
-- [x]  2023-11
-- [x]  2023-10
-- [x]  2023-09
-- [x]  2023-08
-- [x]  2023-07
-- [x]  2023-06
-- [x]  2023-05
-- [x]  2023-04
-- [x]  2023-03
-- [x]  2023-02
-- [x]  2023-01
-- [x]  2022-12
-- [x]  2022-11
-- [x]  2022-10
-- [x]  2022-09
-- [x]  2022-08
-- [x]  2022-07
-- [x]  2022-06
-- [x]  2022-05
-- [x]  2022-04
-- [x]  2022-03
-- [x]  2022-02
-- [x]  2022-01
-- [x]  2021-12
-- [x]  2021-11
-- [x]  2021-10
-- [x]  2021-09
-- [x]  2021-08
-- [x]  2021-07
-- [x]  2021-06
-- [x]  2021-05
-- [x]  2021-04
-- [x]  2021-03
-- [x]  2021-02
-- [x]  2021-01
-- [x]  2020-12
-- [x]  2020-11
-- [x]  2020-10
-- [x]  2020-09
-- [x]  2020-08
-- [x]  2020-07
-- [x]  2020-06
-- [x]  2020-05
-- [x]  2020-04
-- [x]  2020-03
-- [x]  2020-02
-- [x]  2020-01
-- [x]  2019-12
-- [x]  2019-11
-- [x]  2019-10
-- [x]  2019-09
-- [x]  2019-08
-- [x]  2019-07
-- [x]  2019-06
-- [x]  2019-05
-- [x]  2019-04
-- [x]  2019-03
-- [x]  2019-02
-- [x]  2019-01
-- [x]  2018-12
-- [x]  2018-11
-- [x]  2018-10
-- [x]  2018-09
-- [x]  2018-08
-- [x]  2018-07
-- [x]  2018-06
-- [x]  2018-05
-- [x]  2018-04
-- [x]  2018-03
-- [x]  2018-02
-- [x]  2018-01
-- [x]  2017-12
-- [x]  2017-11
-- [x]  2017-10
-- [x]  2017-09
-- [x]  2017-08
-- [x]  2017-07
-- [x]  2017-06
-- [x]  2017-05
-- [x]  2017-04
-- [x]  2017-03
-- [x]  2017-02
-- [x]  2017-01
-- [x]  2016-12
-- [x]  2016-11
-- [x]  2016-10
-- [x]  2016-09
-- [x]  2016-08
-- [x]  2016-07
-- [x]  2016-06
-- [x]  2016-05
-- [x]  2016-04
-- [x]  2016-03
-- [x]  2016-02
-- [x]  2016-01
-- [x]  2015-12
-- [x]  2015-11
-- [x]  2015-10
-- [x]  2015-09
-- [x]  2015-08
-- [x]  2015-07
-- [x]  2015-06
-- [x]  2015-05
-- [x]  2015-04
-- [x]  2015-03
-- [x]  2015-02
-- [x]  2015-01
-- [x]  2014
-- [x]  2013
-- [x]  2012
-- [x]  2011
-- [x]  2010
-- [x]  2009
-- [x]  2008
-- [x]  2007
-- [x]  1969
-- [x]  1947
-- [x]  1904
