--- lightroom-photo-tagging/tsconfig.json
@@ -1,29 +1,0 @@
-{
-  "compilerOptions": {
-    "resolveJsonModule": true,
-    "baseUrl": ".",
-    "paths": {
-      "*": [
-        "types/*"
-      ]
-    },
-    "lib": [
-      "ES2021",
-      "DOM"
-    ],
-    "target": "ES2022",
-    "module": "commonjs",
-    "noImplicitAny": true,
-    "removeComments": true,
-    "preserveConstEnums": true,
-    "sourceMap": true,
-    "strictNullChecks": true,
-    "esModuleInterop": true,
-  },
-  "include": [
-    "src/**/*.ts"
-  ],
-  "exclude": [
-    "node_modules"
-  ]
-}
