+++ portfolio-preparation/.env.sample
@@ -1,0 +1,5 @@
+DEBUG_MODE=false
+# Portfolio metadata generation reads exported .avif photos from here.
+# Must match the Lightroom "For Portfolio - Large" export dir. The blurhash
+# step also expects a sibling dir with "large" swapped for "thumbnail".
+PORTFOLIO_INGEST_PATH='/Users/travisbumgarner/Desktop/large'
