+++ portfolio-preparation/package.json
@@ -1,0 +1,27 @@
+{
+  "name": "portfolio-preparation",
+  "version": "1.0.0",
+  "private": true,
+  "license": "MIT",
+  "scripts": {
+    "start": "ts-node src/main.ts",
+    "lint": "biome check .",
+    "lint:fix": "biome check --write .",
+    "format": "biome format --write .",
+    "precommit": "tsc --noEmit --skipLibCheck"
+  },
+  "dependencies": {
+    "blurhash": "^2.0.5",
+    "date-fns": "^3.6.0",
+    "dotenv": "^16.4.7",
+    "exifr": "^7.1.3",
+    "sharp": "^0.35.3",
+    "uuid": "^11.0.2"
+  },
+  "devDependencies": {
+    "@biomejs/biome": "1.9.4",
+    "@types/node": "^20.12.13",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.4.5"
+  }
+}
