+++ portfolio-preparation/src/config.ts
@@ -1,0 +1,18 @@
+import dotenv from 'dotenv'
+dotenv.config({ path: './.env' })
+
+function notNull(key: string) {
+  const value = process.env[key]
+  if (!value) {
+    throw Error(`Missing environment variable: ${key}`)
+  }
+
+  return value as string
+}
+
+const config = {
+  debugMode: notNull('DEBUG_MODE') === 'true',
+  portfolioIngestPath: notNull('PORTFOLIO_INGEST_PATH')
+}
+
+export default config
